change structur file base and remove unsused file
