+++ src/parts/Header.js
@@ -1,26 +1,16 @@
 import React from 'react';
+import Fade from 'react-reveal/Fade';
+import DarkModeToggle from "react-dark-mode-toggle";
+
 import Button from 'elements/Button';
 import BrandIcon from 'parts/IconText';
-import Fade from 'react-reveal';
+import { isMobile } from 'react-device-detect';
 
-export default function Header(props) {
-  const getNavLinkClass = path => {
-    return props.location.pathname === path ? " active" : "";
+export default function Header({setTheme, theme}) {
+  let RoomId = window.location.pathname.slice('13')
+  const getNavLinkClass = (path) => {
+    return window.location.pathname === path ? " active" : "";
   };
-
-  if (props.isCentered) {
-    return (
-      <Fade>
-        <header className="sapcing-sm text-center mt-2">
-          <div className="container mb-2">
-            <Button className="brand-text-icon" href="" type="link">
-              Stay<span className="text-gray-900">cation.</span>
-            </Button>
-          </div>
-        </header>
-      </Fade>
-    );
-  }
 
   return (
     <Fade>
@@ -28,6 +18,12 @@ export default function Header(props) {
         <div className="container">
           <nav className="navbar navbar-expand-lg navbar-light">
             <BrandIcon />
+            <DarkModeToggle 
+              className={!isMobile && 'ml-3'}
+              onChange={setTheme}
+              checked={theme === 'dark'}
+              size={48}
+            />
             <div className="collapse navbar-collapse">
               <ul className="navbar-nav ml-auto">
                 <li className={`nav-item${getNavLinkClass("/")}`}>
@@ -35,14 +31,19 @@ export default function Header(props) {
                     Home
                   </Button>
                 </li>
-                <li className={`nav-item${getNavLinkClass("/live-stream")}`}>
-                  <Button className="nav-link" type="link" href="/live-stream">
+                <li className={`nav-item${getNavLinkClass(`/live-stream/${RoomId}`)}`}>
+                  <Button className="nav-link" type="link" href="/live-stream/332503">
                     Live Stream
                   </Button>
                 </li>
-                <li className={`nav-item${getNavLinkClass("/agents")}`}>
-                  <Button className="nav-link" type="link" href="/agents">
-                    Agents
+                <li className={`nav-item${getNavLinkClass("/list-room")}`}>
+                  <Button className="nav-link" type="link" href="/list-room">
+                    Room List
+                  </Button>
+                </li>
+                <li className={`nav-item${getNavLinkClass("/about")}`}>
+                  <Button className="nav-link" type="link" href="/">
+                    About
                   </Button>
                 </li>
               </ul>
